requirements.txt
fastapi==0.100.0
uvicorn[standard]==0.23.0
sqlmodel==0.0.8
python-jose==3.3.0
passlib[bcrypt]==1.7.4
pandas==2.2.2
reportlab==4.1.0
pytest==7.4.0
requests==2.31.0
python-dotenv==1.0.0
